+++ packages/fusion/src/chunk.ts
@@ -1,0 +1,11 @@
+/**
+ * Chunk module
+ * @module chunk
+ */
+
+export type { FileChunk } from "#/@types/chunk";
+
+export {
+    isTypeOfFileChunk,
+    isTypeOfFileChunks,
+} from "#/functions/chunk";