+++ packages/fusion-node/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+### What's Changed
+
+- Update in types
+
 ## 0.1.0 (2025-06-09)
 
 First release