+++ packages/fusion/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.2.0
+
+### Breaking Changes
+
+- Move `FileChunk` type:
+    - from: `@filerune/fusion`
+    - to: `@filerune/fusion/chunk`
+- Move `isTypeOfFileChunk` function:
+    - from: `@filerune/fusion`
+    - to: `@filerune/fusion/chunk`
+- Move `isTypeOfFileChunks` function:
+    - from: `@filerune/fusion`
+    - to: `@filerune/fusion/chunk`
+
 ## 0.1.0 (2025-06-09)
 
 First release